feat: add headers for other product types
